requirements.txt
langchain
python-dotenv
ipykernel
langchain_groq
langchain_core
langchain_community
streamlit
faiss-cpu
langchain_huggingface
huggingface_hub
sentence_transformers
pypdf